Prevent problems when saving slides pages
When a slide pack is saved as a complete HTML page, the slides are
already processed and present as HTML elements. So, in order to prevent
a sort of "double render" when the user opens the HTML file, the
markdown container node is marked as "already processed" and will be
skipped.